agent-builder-service/requirements.txt
# Agent Builder Service Requirements

# HTTP client for Kibana API
httpx>=0.25.0
aiohttp>=3.9.0

# Data validation
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
